# api/openapi.yaml
openapi: 3.0.3
info:
  title: CodeForge API
  description: Remote AI code task runner. Submit code tasks, monitor progress via SSE, and create pull requests.
  version: 1.0.0
  contact:
    name: CodeForge
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/workflows:
    post:
      summary: Create a workflow definition
      operationId: createWorkflow
      tags: [Workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDefinition"
      responses:
        "201":
          description: Workflow created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      summary: List workflow definitions
      operationId: listWorkflows
      tags: [Workflows]
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkflowDefinition"
  /api/v1/workflows/{name}:
    get:
      summary: Get a workflow definition
      operationId: getWorkflow
      tags: [Workflows]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinition"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a workflow definition
      operationId: deleteWorkflow
      tags: [Workflows]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/workflows/{name}/run:
    post:
      summary: Run a workflow
      operationId: runWorkflow
      tags: [Workflows]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunWorkflowRequest"
      responses:
        "201":
          description: Workflow run created
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    format: uuid
                  workflow_name:
                    type: string
                  status:
                    type: string
                    example: pending
                  created_at:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Operation conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...
    WorkflowDefinition:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: sentry-fixer
        description:
          type: string
        builtin:
          type: boolean
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepDefinition"
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/ParameterDefinition"
        created_at:
          type: string
          format: date-time
    StepDefinition:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [fetch, task, action]
        config:
          type: object
          description: Step-type-specific configuration
    ParameterDefinition:
      type: object
      required: [name]
      properties:
        name:
          type: string
        required:
          type: boolean
        default:
          type: string
    RunWorkflowRequest:
      type: object
      properties:
        params:
          type: object
          additionalProperties:
            type: string
          description: Workflow input parameters
...